requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
pymysql
python-dotenv
python-multipart
polars
scikit-learn
torch
numpy
matplotlib
